Refactor ExportsCountriesEdit to get all data from resource
