+++ src/apps/companies/apps/exports/views/edit-countries.njk
@@ -1,15 +1,9 @@
-{% extends "_layouts/template.njk" %}
+{% extends "_layouts/template-no-local-header.njk" %}
 
-{% block local_header %}
-  {% call LocalHeader({ heading: 'Edit export countries' }) %}{% endcall %}
-{% endblock %}
-
-{% block body_main_content %}
-
+{% block body %}
   {% component 'react-slot', {
       id: 'company-export-countries-edit',
       props: props
     }
   %}
-
 {% endblock %}
