+++ src/client/modules/Companies/CompanyExports/ExportCountriesEdit/tasks.js
@@ -1,6 +1,6 @@
 import { API_ERROR, API_WARN, SAVED } from './state'
-import urls from '../../../../../../lib/urls'
-import getExportCountries from '../../../../../../lib/get-export-countries'
+import urls from '../../../../../lib/urls'
+import getExportCountries from '../../../../../lib/get-export-countries'
 import axios from 'axios'
 
 function transformFieldValues(fields) {
